Add ec test command that mimics conftest test
For the test coverage I'm recycling rego and data from the existing
acceptance test, rather than creating new stuff, and I'm choosing to
rely on the acceptance tests and to not write any low level unit
tests.

As per the commentary, the new command is deliberately as identical
to the confest version as possible. I think this will make it easier
to keep it up to date over time, by viewing the diff and noting any
fresh changes. The Makefile helper is meant to make it easier to do
that.

JIRA: HACBS-2376