Add new policy-input output format
The policy-input format exports the image data that is provided to the
policy as `input` in the `validate image` command. Given that multiple
images can be validated at the same time, and that ec generates a unique
`input` for each image, the output is in the JSONL format. A policy,
however, only sees one of the objects at a time.

This commit also changes the variables `REGISTRY_*_HASH` to
`REGISTRY_*_DIGEST`. The value does not include the algorithm, i.e.
`sha256`. This is needed because the image-data output contains fields
where only the digest is present, without the algorithm. Having both
variables is problematic because they map to the same value. This means
that whether or not the value is replaced with one vs the other is
not deterministic.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>