Support appstudio formatted output for ec test
Add a function to internal/applicationsnapshot to help prepare it so
we don't have to expose the types outside that file, and a minor
refactor related to calculating the overall result value.

Also added some commentary about the mismatch between the
"namespace" field which is currently a single string instead of a
list of strings.

JIRA: HACBS-2376