Add manifest files from the image to input
This adds manifest files from the image to the input under
`image.manifests`. Files that are located in the `manifests` directory
and can be parsed as JSON or YAML are considered.

Each file is parsed and converted to JSON and included in the input JSON
tree verbatim.

Ref. https://issues.redhat.com/browse/HACBS-2379