Fix development environment
The helm script (`hack/helm.sh`) was failing to compile and run helm due
to dependency issues in a triangle of oras-go, docker and
k8s.io/client-go. This adds a conforming dependency replaces.
Also since we can break the development environment by arbitrarily
breaking tools by updating a dependency a check was added to try to run
the tools that we require and see if they can be compiled and run.