HACBS-2424 Linter for go.mod files.
This commit adds a makefile target that searches for `go.mod` files and
in each directory where one is located, it executes `go mod tidy` to
ensure that the `go.mod` and `go.sum` files are updated. It then
displays a message to the user to remind them to add the modified files
to their git commit.

Signed-off-by: Rob Nester <rnester@redhat.com>