Read and parse atts with non-slsa predicate types
Rather than discarding any attestation that can't be unmarshalled as
SLSA v0.2, now we'll treat them as a potentially useful attestation
that should be visible to the policy checker.

The motivation is to include include SPDX SBOM attestations in the
policy input, such as the ones created with Syft and added to the
registry using cosign attest.

Ref: https://issues.redhat.com/browse/HACBS-2515

Signed-off-by: Simon Baird <sbaird@redhat.com>