Add new map[string]any helper interface
I think this will help make some code that deals with unmarshalled
json cleaner and easier to read. With the unit tests in place I
expect the maintenance burden should be low enough.

I had some fun trying out go generics, which seem to work well and
make the code nicer.

This was created while hacking on HACBS-2515, which is about working
with attestation predicates of an unknown format.