Run the release workflow only for top of HEAD
When the Checks workflow is run and succeeds for a change that targets
the default branch the release workflow will trigger. This includes
reruns and doesn't limit to the last commit on the main branch.

This adds a condition so that the release workflow only proceeds if it
is triggered for a Checks workflow that ran for the latest commit on the
default branch.

Ref. https://issues.redhat.com/browse/HACBS-2585