Fail if unsupported Rego rules are loaded
This adds a check when the Rego files are inspected to see if the policy
rules return a supported value (string, object, variable or a value from
function invocation). This is as fine as we can make this check as there
is no type information from OPA.

Ref. https://issues.redhat.com/browse/RHTAPBUGS-736