Show total checks passed in successes count
Previously the "successes" field in the appstudio output showed just
a count of each component with an overal pass. Change it to show a
count of the number of passing checks.

It seems more intuitive to do it this way, and I think probably the
reason it's like this is because the relevant code pre-dates the
ability to know how many passing checks there were.

Notes:
- To ensure the success count is present even when --show-success
  isn't used, there's an extra field added to the
  applicationsnapshot.Component struct.
- I decide not to include the successCount value in the normal
  report output, hence the `json:"-"` in the struct definition.

Res: https://issues.redhat.com/browse/HACBS-2694