Set fetch depth to 0 for the release workflow
Given a set of versions it is difficult to compare them, for that reason
we introduced the patch version having the number of commits on the
release branch. The fetch depth set to 1 by default sets the patch
version to `1`.

This sets the fetch depth to `0`, i.e. all commits, so that the command
`git rev-list --count HEAD` returns the number of commits on the release
branch instead of `1`.

Resolves: #1111