Workarounds for two build args in RHTAP
I'm not sure how to provide build args it as parameters to the
buildah task. I expect COSIGN_VERSION could be derived while
building, but it's not simple since I'm thinking we'd need to use
curl instead of the ADD command.

For the EC_VERSION, I'm not sure what the long term plan is for
versioning builds, but let's just get something working and figure
out the details later.