Add some labels to the RHTAP docker file
For the first five labels, the motivation is to avoid the EC
violations 'The "foo" label should not be inherited from the parent
image'. See [1] where the list of labels are defined.

I also added a com.redhat.component label. Not sure what it's for
but I thought I'd follow the example from securesign/rekor, see [2].

There might be other labels we want to set in future, including some
others that are inherited from the base image like name and
build-date for example.

[1] https://github.com/release-engineering/rhtap-ec-policy/blob/b56ae2237264f7f196a9e6c606d5cdb86948312b/data/rule_data.yml#L94
[2] https://github.com/securesign/rekor/blob/c43d979f46215e1069d101e9dae96cdcbf42cad9/Dockerfile.cli#L30