Include an increasing build number in full version
The motivation for this is to make it easier to know if in ec build
is newer than another ec build.

Requires changes to the git-clone task params so we have enough git
history to choose the build number.

The number we're getting now is quite high (2387), but if we add a
tag at some point it will shrink down to a more sensible number.