Add a comment about local build gotcha
I considered different ways of "fixing" the problem, e.g. by
cleaning out the dist directory in the docker file, or by creating
the binaries in a different directory, but in the end I felt like it
wasn't worth the extra complexity, so here's a comment to warn about
it instead.