Script for applying RHTAP pipeline changes
Since each release requires a new RHTAP application and a new
pipeline definition, we need a way to apply the required pipeline
changes for each release in the branch for that release.

This commit adds some bash scripting to do that. It's expected the
script will be run by a human so they can review the result manually.

Note: If we find the technique of scraping the patches from the commit
history becomes too inflexible we should check in some suitable patch
files (under hack/patches maybe) and use those instead.

Ref: https://issues.redhat.com/browse/EC-342