Use git for-each-ref to detect branch name
We're using the branch name to derive the release name, but it
doesn't work in the TAP build pipeline because the git-clone task
checks out the sha directly and not the branch.

With this patch it should find the branch name correctly anyway.
We're making some assumptions: there is only one branch name that
points to the HEAD sha, and the git-clone task fetched down the
branch refs from origin, which does seem to be happening as a
side-effect of fetching the tags. (See aso the fetchTags param
in the git-clone task which we set to "true".)

Ref: https://issues.redhat.com/browse/EC-343