Use generic name if there's no branch name version
I don't think there's any significant impact for this, but for
cosmetic reasons I think it's nicer to use something other than an
empty string.

Also fix a bug causing all arch/os builds to be build for a
per-merge PR build.

Note: The bash logic in the Docker file is starting to look quite
messy. We might consider moving it to the Makefile if it gets any
busier.