Better go mod download layer caching
It probably makes no difference when building in Konflux (formerly
RHTAP), but for local builds, e.g. when you're testing and debugging
the Dockerfile, this can save some time.

It works because podman can now decide it doesn't have to rebuild
the layer with all the go modules if the mod file didn't change from
last time it was built.