wip: Introduce the concept of a validation intent
This commit adds a new flag to the `ec validate command`, `--intent`
which is propagated to rego as config value - similarly to how
effectiveTime is propagated.

The idea is to provide a mechanism to change the behavior of a policy
rule depending on why the user is validating an image.

For example, the
[schedule](https://enterprisecontract.dev/docs/ec-policies/release_policy.html#schedule_package)
package contains policy rules that are meant to enforce restrictions
depending on the day. This is meant to disallow releases on the weekends
or on certain days (holidays).

Those policy rules are NOT meant to disallow changes from being merged.

This allows the user to use the same policy config for multiple
purposes.

Ok, so why not just have multiple policy configs?

This is certainly a possibility! It does come with its own challenges:

1. Deviation in what is checked during integration vs release is prone
   to cause unexpected issues at release time.
1. Maintenance. Duplicating every policy config sounds like a good way
   to increase toil for everyone.

The "intent" appraoch does away with the maintenance cost for users, at
the expense of making certain policy rules slightly more complex. (I
think that's a good trade-off.)

But it only reduces the deviation problem. You'd still get a different
result in integration than in release. But... that's what the requested
feature is all about so not sure if we can do much about that. This
approach does use the same policy config in both cases so at least it
removes issues caused by those not being in sync.

Anyways, this is a proof of concept meant to help the conversation.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>