Support adding only digest bumps to pipeline
If the pipeline was created by RHTAP/Konflux a while ago, it's
likely they task bundle digests are old. This provides a way to
sync the the bundle digests with main branch.

There are probably nicer ways to do this, and I'm not sure all
patches will apply cleanly in the future (e.g. for out next release)
but let's move quickly and figure it out later.