Add dockerfile for tas client server
It's a long story, but this is an alternative workaround for the
problem where images build from
https://github.com/securesign/sigstore-ocp/blob/9024422e356c22660af7870aecd55a0571f07f58/images/Dockerfile-clientserver
have too many base images, so they overflow the tekton task result
that lists all the base images.

Just a POC so far but it works for me locally. Not sure what the
next step would be, but we could make a component and try building
this in Konflux. And we need an updated base image since the one I'm
using here is very old and doesn' have all the different arches.