Add ec.oci.image_manifest rego function
This commit adds a new custom rego function, `ec.oci.image_manifest`.
This function retrieves the Image Manifest from an OCI registry. (It
does not download the image, just its manifest.)

The main use case this is trying to achieve is validating image
references that may occur in the SLSA Provenance attestation of an image
being validated. For example, the SLSA Provenance may contain a link to
the corresponding source container image. This function allows policy
rules to be created to verify that such references actually exists.

Ref: EC-235

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>