Keep patching pipelines if one patch doesn't apply
When using the patch pipeline script for the v0.2 release branch,
one of the patches, specifically the change from f1a1972, didn't
apply. The patch program interactively asked by first if I should
"reverse" it (no), then if I should apply it anyway (no).

The behavior is what we want, but it returns an error code, so the
script aborts. Make it so the script keeps going in that kind of
situation.

It does mean you have to pay attention and look at why and how the
patch doesn't apply.

Ref: https://issues.redhat.com/browse/EC-434