Use explicit non-root user in Konflux image build
Specifying with a uid since because we're not sure that the 1001
user exists in /etc/passwd, and we get errors otherwise.

(To expand on that a little - there is a "default" user with uid
1001 that *does* exist in registry.access.redhat.com/ubi9/go-toolset:latest,
but it does not exist in the brew.registry.redhat.io/rh-osbs/openshift-golang-builder
base image we're using currently.)

There was some discussion about the potential side-effects of this
change, particularly if volumes need to be mounted. I'm not 100%
confident about it, but maybe it's fine.

The goal, once again, is to get the OpenShift preflight check task
to stop producing failures.

Ref: https://issues.redhat.com/browse/EC-464