Add script to copy the Konflux task pipeline
The new task bundle Konflux pipeline is not generated initially by
Konflux. Instead it was created manually (IIUC using Kustomize with
the template in the build-definitions repo, ask Joe for the details
if you want to know more).

So to create it in the release branch we can just copy it from main
branch and make a few adjustments to ensure it is linked to the
correct Konflux component.

https://issues.redhat.com/browse/EC-428
https://issues.redhat.com/browse/EC-472