EC-30 Add check for policy conformance to schema
This commit adds a check to see if the provided policy configuration is
conformant to the Enterprise Contract Policy spec schema.

In order to ensure that a provided policy configuration is conformant to
the Enterprise Contract Policy Spec schema for the version of
EnterpriseContractPolicySpec imported as a dependency, we generate the
schema at compile time in lieu of referencing an external version.

Signed-off-by: robnester-rh <rnester@redhat.com>