Don't rebuild task bundle on every commit
Only build it if the relevant yaml file changed.

Note: We would want to trigger build also if the pipeline definition
changed. I'm not sure if we need to explicitly list
.tekton/verify-enterprise-contract-task-main-ci-pu*.yaml in the
expression, or whether that would be implied automatically, but we
should be able to find out once a dependabot digest bump PR gets
merged. If it triggers a build we'll know we don't need to
explicitly mention it in the expression, and vice versa.

Ref: https://issues.redhat.com/browse/EC-476