Replace vith `${TIMESTAMP}` only for current dates
Dates with `0` on the second digit will not be replaced, this leaves
dates that are significant, i.e. those in the past or in the far future
within the snapshot.
Now we can see these significant dates in the snapshots, i.e. we can see
that an `effective_on` date is kept the same.