Support for mocking builtin functions
Adds a new builtin: `ec.testing.mock` that can be used only when invoked
via `ec opa test` command. When invoked with a name of the function,
arguments and the mocked return value, that information is stored in the
private static map within the `internal/rego/testing` package, and
supporting builtins can check for a short circuit mocked value return by
invoking the `testing.Mocked` function.