Simpler version number logic for release branch
This is a quick-fix, and partly motivated by EC-511. I'm having
trouble getting a green build in release branch, and I'm hoping this
will make a re-triggered build behave sanely.

Remove the complicated logic related to detecting the branch name,
extracting the version number, and deciding if it's a real release
build. Just hard code it. (Will worry about porting this to main
branch later.)

Ref: https://issues.redhat.com/browse/EC-511