Generate reference documentation at build time
This replaces the use of
`@enterprise-contract/reference-antora-extension` with code generation
that is run at build time. The documentation should resemble the
existing documentation generated at website build time, only the point
at which the documentation is generated has changed. Also the
documentation is now versioned in the git repository, so a check was
added to make sure the documentation is up to date, i.e. doesn't need to
be regenerated.

Reference: EC-513