Add policy.yaml for ec verifications in Konflux
Now we have a need to use a customized policy so let's create one. I
figured we could keep it here in this repo under .ec, in the style
of .tekton, .github, etc.

As mentioned in the comments, ec by convention will look for a
.ec/policy.yaml file at the git url specified.

Right now it's effectively the same as redhat-non-hermetic, but see
next commit for the newly required customization.

See here for the original redhat-non-hermetic:
https://github.com/enterprise-contract/config/blob/main/redhat-no-hermetic/policy.yaml

Ref: https://issues.redhat.com/browse/EC-520