EC-417 Refactor to stop repeated source downloads
This commit refactors the creation of the confestEvaluator so that they
are created once, one per source group, instead of once per source group
per component. This should dramatically reduce the number of
confestEvaluators created for applications with many components.

This commit also refactors the GetPolicy method to cache downloaded
sources in a given run, avoiding the need to download them repeatedly.

Signed-off-by: robnester-rh <rnester@redhat.com>