Customize the newly created task build pipeline
The changes here are based on the pipelines that were
removed in #1516 when the verify-enterprise-contract-task-main-ci
component was removed from ec-main-ci.

The main signficant changes is to use tkn-bundle instead of buildah
as the build task.

There are also some tasks being removed. We could try bring them
back later, but for now I just want the shortest path to getting
something working.

Background: The task definition build was permanently failing its EC
check leaving all our PRs with failing CI. The plan here is to bring
it back but in a different Konflux application, perhaps with no EC
check, or an EC check with a config customized so that it can be
passing.

Note that the task result JAVA_COMMUNITY_DEPENDENCIES is created by
the tkn-bundle task, so that's why it's removed from the pipeline
results.