Ensure evaluators are not destroyed if reused
We changed the life cycle of evaluators to the top of the command, and
mistakenly left the destruction of the evaluator tied to the lifecycle
of validating a single component.
This causes files created for one evalator and shared between component
validations or evaluators inaccessible. An error such as:

```
get compiler options: capabilities not opened: open /tmp/ec-work-299517100/capabilities.json: no such file or directory
```

would be reported.

This changes the destruction of the evaluator to be in line with the new
lifecycle.