Add Makefile task to bump minor version
It increments the minor version number and creates a commit. I think
it's nice to have a consistent way to do this, even though it's not
hard to do it manually, as mentioned in the comments.

This is expected to be run by a human locally, not in automation
workflows.

(Will use this to create the next commit.)

Ref: https://issues.redhat.com/browse/EC-522