Enable prefetch task in main branch Konflux builds
Note that the default pipeline skips prefetch unless the hermetic
param is set to true. To make prefetch run without hermetic being
set, I removed the when clause entirely.

(I did also consider changing the when clause to check for non-empty
prefetch-input. I think that idea is useful for the default pipeline
definition in build-definitions repo, but it doesn't do much for us
here, so I decided for this PR that we should keep it simple.)

Ref: https://issues.redhat.com/browse/EC-585