Set snapshot release to "latest"
See relevant docs at
https://github.com/softprops/action-gh-release?tab=readme-ov-file#-customizing

It's necessary to remove the "prerelease" option since a prerelease
build can't be designated as the latest.

(I was hoping to prevent the rolling release from being pushed down
the page every time a new release is created, but it looks like
that's still happening. It seems reasonable to designate the rolling
release as the latest, so let's keep this change anyhow.)

Ref: https://issues.redhat.com/browse/EC-602