Use Sigstore staging for test recordings
This commit modifies the acceptance tests which rely on recordings to be
based on the staging instance of Sigstore.

The main advantage of this approach is that to regenerate the data, we
no longer need to spin up a local instance of Sigstore which can be
quite resource intensive. Also, since the stagning Sigstore instance is
not an ephemeral instance, we can be selective in only re-generating the
data that expires.

It would be great to eventually tie this into a GitHub Workflow so a PR
can be created every so often.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>