Use go-toolset golang base image for redhat builds
The --chown=1001:0 was required to avoid an error like this:

  fatal: detected dubious ownership in repository at '/build'
  To add an exception for this directory, call:
          git config --global --add safe.directory /build

I'm not exactly sure the reason why, but I believe it's because the
new base image comes with a user called 'default' with uid 1001.
When we copy the source code into the build container it comes in
owned by root, and that causes git to be suspicious when it runs as
user default.