Fix logic for no matching rule
If a source group produces no warnings, failures, or successes, we want
to raise an error as that is a no-op and likely indicates a typo in the
policy configuration.

This commit fix the check by ensuring the total number of matching rules
is computed *after* the exclude/include from the policy config is
applied.

Ref: EC-583

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>