Less all-caps constants for output formats
IIUC the convention is to use all-caps for acronyms, so I think we
have a little too much of it here.

I also changed the order around. It's not consequential particularly
but I wanted to match the order in the list from the previous commit
for the sake of tidiness. (The order of that list does impact the
command line help for `ec validate image`.)

Also tweak the switch statement for the deprecated hacbs output.

Ref: https://issues.redhat.com/browse/EC-563