Sync params between push and pull Pipelines
This commit adds two parameters to the embedded push build Pipeline,
`image-expires-after` and `build-args-file`.

These are used by the pipelineSpec of both push and pull Pipelines. In
the push case, those values are empty and effectively a no-op. They are
added to both Pipelines simply to facilitate keeping those Pipelines in
sync.

Ref: EC-556

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>