Set k8s version for `helm template`
In the `hack/setup*` scripts we use `helm template` indirectly via
`kustomize`. The version of Kubernetes in that case is hardcoded to
`1.20.x`, which is causing a version check from the ingress chart to
fail. This sets the k8s version to the version that is read at runtime
from the launched kind cluster.