Don't wrap image ref in text output
The wrapping only wraps on whitespace, so before this patch we'd end
up with output like this for long image refs, which IMO looks
untidy. Let's just have an over-long line rather than wrapping it
like that.

```
› [Warning] cve.unpatched_cve_warnings
  ImageRef:
  quay.io/redhat-user-workloads/somelong-app-name/some-long-component-name@etc...
```

There are a few ways to do this it, but I felt like the indent
helper works best with the agressive space eating done by {{- and -}},
and keeps it consistent with the other things we're outputing.

This is a followup tweak for the functionality introduced in #1656.