Handle Index Images with no-platoform manifests
When expending an Image Index into Image Manifests, a component is
expanded into pseudo components. These are suffixed with the
architecture of the particular Image Manifest. However! The manifest
descriptor in an Image Index are not required to provide a platform
(where the architecture is defined).

Prior to this commit, such Image Indexes would cause a segfault due to a
nil pointer exception. This commit resolves that issue by defaulting to
a different value if a platform is not specified.

Fixes #1671
Ref: EC-674

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>