chore: Modify default cli pipelines for v0.4
Apply changes to the Konflux generated default pipelines.

Also remove the main branch pipelines since they're not needed in the release branch.

Commit initially created with hack/patch-release-pipelines.sh, but
there was a lot of manual updates since the -oci-ta hunks didn't
apply cleanly for some reason.

Vimdiff was used like this:
  vimdiff +'set ft=yaml' <(git show main:.tekton/cli-main-ci-pull-request.yaml) .tekton/cli-v04-pull-request.yaml
  vimdiff +'set ft=yaml' <(git show main:.tekton/cli-main-ci-push.yaml) .tekton/cli-v04-push.yaml
and all the changes that looked right were applied manually by
copying from the left and pasting on the right.

Notes:
* We're converting from the standard tasks to the '-oci-ta' tasks in
  this commit.
* Ideally we should switch to the quay.io/konflux-ci/tekton-catalog
  repo instead of the older repo, but to minimize the amount of
  change I'm not making that change yet.
* There are some things visible in the diff that are new in the
  default Konflux pipeline that we should back-port to the main
  branch pipeline, but that will be addressed later.