Use pinned policy refs in acceptance tests
The goal is to get the acceptance tests to pass in this old branch
so I can merge this PR.

It's likely they're failing because of the use of the `ec.oci.*` and
maybe also `ec.sigstore.*` custom internal rego functions in the
rego from ec-policies. Those functions are not defined in this older
version of ec.

The release-v0.1-alpha branch of ec-policies was created just now,
see https://issues.redhat.com/browse/EC-695 . It points to the
commit right before those custom functions were introduced.

Note:
- This is a quick hack in a branch that's abandoned. We might want
  to create a nicer and more generalized version of this in the current
  main branch so the acceptance tests don't always assume they can use
  main branch of ec-policies.
- For the longer term: It does seems a little odd to be reaching out
  to another repo while running tests in this repo. We might want to
  rethink that at some point, but I suspect that won't be for a
  while.