Generate RPM lock file
The `hack/update-rpm-lock.sh` script will run the
`rpm-lockfile-prototype`[1] tool configured with the `Dockerfile.dist`
and `rpms.in.yaml` files to generate the `rpms.lock.yaml` file. The lock
file can be read by cachi2[2] to precache RPM packages with pinned
versions. And that in turn enables hermetic builds.

`hack/update-rpm-lock.sh` script helps by installing the needed tools
and dependencies in a container image that is later run to execute the
`rpm-lockfile-prototype` tools. The script also extracts any `.repo`
files in `/etc/yum.repos.d` which can be referenced in the
`rpms.in.yaml` file.

Reference: EC-360

[1] https://github.com/konflux-ci/rpm-lockfile-prototype
[2] https://github.com/containerbuildsystem/cachi2/