Humanize v-e-c Task logs
The top of the logfile for a TaskRun of verify-enterprise-contract Task
should be the text (human readable) report. Following it are the YAML
and JSON logs and preceded by a break `----- DEBUG OUTPUT -----` are
the debug logs and version information.
This is to make the TaskRun's logs easier to understand by humans.

Reference: EC-564