Per output format options
In addition to the global flag `--show-successes` that influences all
outputs and this adds `--show-successes` equivalent per output format.
This allows the full report with successes to be present in JSON or YAML
outputs while the output in text format can be outputted without
successes.

This is done by appending options in URL query format to the output
format following a question mark, e.g.
`--output text=output.txt?show-successes=false`.

Reference: EC-564