KFLUXBUGS-1505 Fix EC OOM by specifying workers
This commit attempts to resolve an issue resulting in ec OOM when
attempting to validate images with large SBOMs by providing a flag to
limit the number of parallel works to reduce memory usage.

The verify-enterprise-contract task has also been updated to utilize
this parameter, with an associated default of "1" workers.

Signed-off-by: robnester-rh <rnester@redhat.com>