Add CPU and memory profiling when tracing
When trace is enabled CPU and memory profile will be written to
`cpuprofile.*` and `memprofile.*` files, these can be later collected
and analyzed by `go tool pprof`.

For this reliable way of witting the profile data was needed, and as it
turns out `PersistentPostRun` is not invoked when error occurs, so an
additional `OnExit` function was added for doing this.

Reference: EC-733