Change default validate image output to text
The previous default was json.

There was one change in the unit tests required to make the tests
pass. I'm not 100% sure why that was needed (or rather why it wasn't
needed before), but my understanding is that the fs and the ctx was
being reused between two test cases and it was causing problems
somehow.

I did look at the task definition also, but the output is specified
explicitly, so should be no change needed.

Ref: https://issues.redhat.com/browse/EC-758