Add validate image default output test coverage
Motivated by CodeCov complaining about one uncovered code path.

I decided to test the output with and without --show-successes. It's
perhaps more than was needed, but actually this is the first unit
test coverage of the validate image text output, (existing coverage
is all in the acceptance tests), so I think it's worthwhile.

Ref: https://issues.redhat.com/browse/EC-758