Do cosign initialize using ec in Konflux image
Add a bash script wrapper for ec sigstore initialize that allows a
"cosign initialize" command to work.

It's a long story, but we're trying to remove the cosign binary from
the image. It's done already in main branch, but removing it from
this image in the release branch makes me worry about RHTAP users
since I'm not confident when or how they'd get an upgraded version
of the tekton task.

The task used by RHTAP users is in the tssc-sample-pipelines repo,
see [1] and [2], so we could update that, but it's possible that the
RHTAP users forked that repo, or copied the task and customized it,
etc. So there would be a good chance we'd break them if cosign was
removed.

Also changed the task definition back to "cosign initialize". Either
would work, but this way our test suite should confirm the
backwards compatibility wrapper script works as expected.

Ref: https://issues.redhat.com/browse/EC-584
Ref: https://issues.redhat.com/browse/EC-817

[1] https://github.com/redhat-appstudio/tssc-sample-pipelines/blob/822d9a01d031ceb772bd2cca0fc7495ad91ac3c1/pac/gitops-repo/gitops-on-pull-request.yaml#L11
[2] https://github.com/redhat-appstudio/tssc-sample-pipelines/blob/822d9a01d031ceb772bd2cca0fc7495ad91ac3c1/pac/tasks/verify-enterprise-contract.yaml#L107

(cherry picked from commit 6f11f6a7bffbe67890a2c9b4efb76dada1f7b810)