Don't produce error output for invalid purl checks
I noticed the error output in when running the unit test in
https://github.com/enterprise-contract/ec-policies/pull/1126

It seems like this should not be an error. Since the purpose of this
method is to decide if the purl is valid, quietly returning false if it
is invalid would be all that it should do. Logging an error seems
unexpected.

(For the parse functon I didn't change it. My reasoning is that if
you call parse then you're expecting it to be parseable. If it fails
at that point then it's unexpected, so logging an error does seem
reasonable in that case.)

Tangentially related to...
Ref: https://issues.redhat.com/browse/EC-847